+++ account-abstraction/src/user_actions/mod.rs
@@ -1,3 +1,1 @@
-pub mod common_types;
-pub mod custom_callbacks;
 pub mod execution;
